refactor(pipeline)!: move TrafficImage to data.py to break circular import
