+++ pipeline/data.py
@@ -6,9 +6,14 @@
 
 import json
 from datetime import datetime
+from pathlib import Path
+from typing import Optional
 
 from pydantic import BaseModel
 
+from data import Camera, Rating
+from model import Model
+from timetools import datetime_sgt
 
 class Location(BaseModel):
     """Geolocation consisting of longitude and latitude."""
@@ -23,6 +28,19 @@ class Rating(BaseModel):
     rated_on: datetime
     model_id: str
     value: float
+
+    @classmethod
+    def from_traffic_image(cls, image: TrafficImage) -> "Rating":
+        if image.processed_on is None or image.congestion_rating is None:
+            raise ValueError(
+                "Invalid TrafficImage: Either 'processed_on' or 'congestion_rating' is None."
+            )
+
+        return cls(
+            rated_on=image.processed_on,
+            value=image.congestion_rating,
+            model_id=Model.MODEL_ID,
+        )
 
 
 class Camera(BaseModel):
@@ -42,7 +60,73 @@ class Congestion(BaseModel):
     rating: Rating
     updated_on: datetime
 
+    @classmethod
+    def from_traffic_image(
+        cls, image: TrafficImage, camera: Camera, updated_on: datetime
+    ) -> "Congestion":
+        return cls(
+            camera=camera,
+            rating=Rating.from_traffic_image(image),
+            updated_on=updated_on,
+        )
+
 
 def to_json_dict(model: BaseModel):
     """Convert given pydantic model into the its JSON dict representation"""
     return json.loads(model.model_dump_json())
+
+
+class TrafficImage:
+    """Traffic Image to be rated for congestion.
+
+    Attributes:
+        image: URL that retrieves Traffic camera image.
+        processed: Whether this TrafficImage instance has been processed.
+        congestion_rating: 0-1 Congestion Rating
+        camera_id: ID of the camera that captured this image
+        longitude: Longitude of the camera that captured this image
+        latitude: Latitude of the camera that captured this image
+        processed_on: Timestamp when this TrafficImage instance has been processed.
+    """
+
+    def __init__(
+        self,
+        image,
+        camera_id,
+        longitude,
+        latitude,
+        processed=False,
+        congestion_rating=None,
+        processed_on: Optional[datetime] = None,
+    ):
+        self.image = image
+        self.processed = processed
+        self.congestion_rating = congestion_rating
+        self.camera_id = camera_id
+        self.longitude = longitude
+        self.latitude = latitude
+        self.processed_on = processed_on
+
+    @classmethod
+    def from_camera(cls, camera: Camera, image_path: Path) -> "TrafficImage":
+        """Create TrafficImage from camera & its image path.
+
+        Args:
+            camera:
+                Camera model to create TrafficImage from.
+            image_path:
+                Path to the retrieved captured image from the Camera.
+        Returns:
+            Constructed TrafficImage
+        """
+        return cls(
+            image=str(image_path),
+            camera_id=camera.id,
+            longitude=camera.location.longitude,
+            latitude=camera.location.latitude,
+        )
+
+    def set_processed(self, congestion_rating):
+        self.processed = True
+        self.congestion_rating = congestion_rating
+        self.processed_on = datetime_sgt()
